Grappling gun - hookshot implementation
This commit provides the hookshot mechanic for the grappling gun, this has made use of a state machine as the functions needed to be in the update function but not run along side each other. This mechanic has been created much the same as the grappling swing, although, this needed to take momentum into account so then the player would continue along the path while falling if the hookshot is cancelled hence the movement calculations are different. **NOTE** the movement/ camera control scripts along side half of the grappling swing script were developed in a test environment on a previous day then brought into this project, all mechanics will be developed in this project from now on.