Wind Zone
The wind zone physics have been implemented with this commit, this uses a box collider which acts as the wind effect zone using on trigger and on exit to determine when to apply force to the object, the calculation uses a vector 3 to get the direction and then * by a given strength. Next the wall jump physics will be implemented.