Wall Run/Jump Implementation
A mid development decision has been made to include a wall run feature as I feel that it would complement the other mechanics. This has been done using raycasts to detect if the wall is on the right or the left side of the player so that the correct force direction can be determined, some addition raycasts have been included so that to stop some errors which occurred during the development of this code such as the player being able to fly up the wall when moving and wall jumping. A bool function has also been added to determine if the player is in a position to wall jump bu shooting a raycast down to check if there is ground below if there is not the player can begin a wall run.