Falling Floors/ Platforms
This commit is the implementation of the false floors/ platforms. These work by using raycasts to detect if the object is tagged with "fake platform" and if it is it was turn the gravity of the rigidbody on and unfreeze the constraints while adding a downwards force to give the appearance of the platform/ floor giving way. This works for the grapple swing, hookshot and when a false floor is stepped on, the player will get a second before the platform falls to give opportunity to move some kind of indication they have stepped onto a fake platform will be implemented later. If the player is mid swing or hookshot it will cancel them once the platform gives way causing them to fall.