Meteorite Logic
This commit is to add the meteors which fall from the sky. If the player is within the radius of the meteor when it lands it will knock the player backwards while sending them upwards, this will be used to add an additional challenge to the gameplay. This has been done by setting up an array of transforms and having a random one selected to spawn each instance this happened every second and also adds a random offset to the position of the meteorites this allows for variance as to where the meteorite will land as if they were landing in the same place every time the player could just learn where they fall and avoid them. Once the meteor has hit the ground it enters an on trigger event which applies an explosion force to the rigidbody of the player which causes the knockback effect and it then destroys itself 1 second after impact, visual/ audio effects will be added later.