Environment Build
This commit was the overall build of the levels environment, the level now has a start and end point and is playable but it still needs some additional touches adding to it. Some of the mechanics have been altered and finalized also such as the hook shot mechanic working on ammo.