Visual Additions/ Pause Menu UI
A rope visual has been created on the line renderer, also the hookshot ammo now displays on screen along with a timer so that the player can track their best time. In addition to this a pause menu has been created using text mesh pro for the text displays, this UI is now fully functional and offers the player the options to resume, check the game controls, restart the game, or quit.