Level Detail/ Ammo Positioning
These changes cover adding detail to the level to make it more alive while also providing hints to the kinds of things the player will need to deal with i.e. wind animations on trees in the wind areas. This commit also adds the ammo locations of the hooks which the player will need to pick up in order to complete the level, the level has been fully played through and can be completed a win screen will be added later.