Addition of Audio/ Finalizing Environment
Audio has now been added into the game in the form of sound effects and background music, the settings of these audio sources have been manipulated in such a way that they play in tandem with each other and do not overpower the sound being played by another source. The environment has also been finalized in that the visual wind particle effect has been placed all wind zone areas, this along with a couple of other environmental fixes means the game is now complete with the final requirements being a main menu and a win screen which will be worked on next.